Trying to update to tf2.0.

+++ nalp/datasets/one_hot.py
@@ -119,8 +119,8 @@ class OneHot(Dataset):
             targets.append(tokens_idx[i+max_length])
 
         # Creates empty numpy boolean arrays for holding X and Y
-        X = np.zeros((len(inputs), max_length, vocab_size), dtype=np.int32)
-        Y = np.zeros((len(inputs), vocab_size), dtype=np.int32)
+        X = np.zeros((len(inputs), max_length, vocab_size), dtype=np.float32)
+        Y = np.zeros((len(inputs), vocab_size), dtype=np.float32)
 
         # Iterates through all inputs
         for i, input in enumerate(inputs):
